refactor(tanstack-start): extract server-fn logic to the adapter, make test apps standalone
Moves the admin-page render, server-function dispatch/serialization, layout
data, and API-route handling out of the shippable TanStack app and into
`@payloadcms/tanstack-start` (`loadAdminPage`, `createServerFunctionClient` +
`stripUnserializable`, `loadLayoutData`, `handleAPIRoute`). The app keeps only
thin `createServerFn` shims that supply the app-owned `config` and generated
`importMap`, colocated under `app/_payload`.

Also makes each test suite's TanStack app completely standalone: per-suite
`router.tsx`, routes, components, shims and generated `importMap.js` /
`routeTree.gen.ts`, selected via `srcDirectory`. No file reaches into the root
app; the `~` alias is removed in favor of relative imports.
