+++ .gitignore
@@ -367,6 +367,10 @@ test/payload.db
 
 # TanStack Router auto-generated route tree
 app-tanstack/routeTree.gen.ts
+test/*/app-tanstack/routeTree.gen.ts
+
+# Per-suite standalone TanStack apps generate their own importMap at test time
+test/*/app-tanstack/importMap.js
 
 # Screenshots created by Playwright MCP
 .playwright-mcp
