+++ app-tanstack/app/__root.tsx
@@ -1,7 +1,7 @@
 import { withPayloadRoot } from '@payloadcms/tanstack-start/client'
 import { createRootRoute, HeadContent, Scripts } from '@tanstack/react-router'
 
-import { HydrationMarker } from '~/components/HydrationMarker/index.js'
+import { HydrationMarker } from '../components/HydrationMarker/index.js'
 
 export const Route = createRootRoute({
   head: () => ({
@@ -24,10 +24,10 @@ export const Route = createRootRoute({
   // Single Payload integration touch point: `withPayloadRoot` renders the
   // Payload admin document shell on `/admin` routes and our own shell
   // everywhere else. No root loader, no manual theme/html threading.
-  shellComponent: withPayloadRoot(MarketingHtml),
+  shellComponent: withPayloadRoot(MarketingRoot),
 })
 
-function MarketingHtml({ children }: { children: React.ReactNode }) {
+function MarketingRoot({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en" suppressHydrationWarning>
       <head>
