+++ app-tanstack/app/_payload/api.$.ts
@@ -1,8 +1,9 @@
 import { createFileRoute } from '@tanstack/react-router'
 
 const handler = async ({ request }: { request: Request }) => {
-  const { handleAPIRoute } = await import('~/functions/handleAPIRoute.js')
-  return handleAPIRoute(request)
+  const { handleAPIRoute } = await import('@payloadcms/tanstack-start/server')
+  const config = (await import('@payload-config')).default
+  return handleAPIRoute({ config, request })
 }
 
 export const Route = createFileRoute('/_payload/api/$')({
