+++ app-tanstack/tsconfig.json
@@ -7,8 +7,7 @@
     "rootDir": "..",
     "baseUrl": ".",
     "paths": {
-      "@payload-config": ["../test/_community/config.ts"],
-      "~/*": ["./*"]
+      "@payload-config": ["../test/_community/config.ts"]
     }
   },
   "include": [
