+++ app-tanstack/vite.config.ts
@@ -48,28 +48,22 @@ const port = Number(process.env.PORT) || 3000
 
 const testSuite = process.env.PAYLOAD_TEST_SUITE || '_community'
 
-// Most test suites only exercise the shared admin shell, so they fall back to
-// the shippable routes in `app`. A suite that needs extra front-end routes
-// (e.g. `live-preview`, `admin-bar`) ships its own complete routes directory
-// under `test/<suite>/app-tanstack/app`, keeping that test-only routing out of
-// the shippable root. Mirrors how each Next test suite owns its own `app` dir.
-// An absolute `routesDirectory` overrides the default resolution; this app sets
-// `srcDirectory: '.'` so routes/entries/`routeTree.gen.ts` live at the app root
-// (no `src` wrapper), matching the Next `app` dir layout.
-const suiteRoutesDir = path.resolve(__dirname, '..', 'test', testSuite, 'app-tanstack', 'app')
-const routesDirectory = fs.existsSync(suiteRoutesDir) ? suiteRoutesDir : 'app'
+// Each test suite that ships its own TanStack app
+// (`test/<suite>/app-tanstack`) is a completely standalone app — its own
+// `router.tsx`, routes, `importMap.js`, components and `routeTree.gen.ts`. We
+// point `srcDirectory` at that suite dir so everything resolves within it; the
+// rest fall back to the shippable root app. Mirrors how each Next test suite
+// owns its own `rootDir`/`app` dir. There is no `src` wrapper, so the source
+// dir is the app root itself (`.` for the shippable app).
+// `srcDirectory` must be relative to this app root: TanStack's entry planner
+// joins it onto the root (`path.join`), so an absolute path would be
+// concatenated rather than replace it.
+const suiteDir = path.resolve(__dirname, '..', 'test', testSuite, 'app-tanstack')
+const srcDirectory = fs.existsSync(suiteDir) ? path.relative(__dirname, suiteDir) : '.'
 
 export default defineConfig((env) =>
   mergeConfig(
     payloadPlugin({
-      // Resolve the `~` → app-root alias as an explicit Vite alias rather than
-      // relying on `resolve.tsconfigPaths`. Per-suite route dirs (e.g.
-      // `test/<suite>/app-tanstack/app`) live outside this project, so their
-      // nearest tsconfig is not `app-tanstack/tsconfig.json` and tsconfck would never
-      // apply the `~/*` mapping to them. A Vite alias applies globally, so the
-      // duplicated shell route files resolve shared modules the same wherever
-      // they live.
-      additionalAliases: [{ find: '~', replacement: path.resolve(__dirname) }],
       additionalIgnoreImporters: [
         /^\.\.\/packages\/tanstack-start\/src\/views\/AdminView\.tsx(?:\?.*)?$/,
       ],
@@ -78,11 +72,9 @@ export default defineConfig((env) =>
         exclude: [],
         include: /\.[jt]sx?$/,
       }),
-      routesDirectory,
-      // No `src` wrapper — routes, `router.tsx`, and `importMap.js` sit at the
-      // app root, so the TanStack source dir is the project root itself.
+      routesDirectory: 'app',
       rscPlugin: rsc({ serverHandler: false }),
-      srcDirectory: '.',
+      srcDirectory,
       tanstackStart,
     })(env),
     {
