+++ packages/tanstack-start/src/exports/layouts.ts
@@ -2,3 +2,4 @@ export { getLayoutData } from '../layouts/Root/getLayoutData.js'
 export type { GetLayoutDataArgs } from '../layouts/Root/getLayoutData.js'
 export { RootLayout } from '../layouts/Root/index.js'
 export type { RootLayoutData, RootLayoutProps } from '../layouts/Root/index.js'
+export { loadLayoutData } from '../utilities/loadLayoutData.js'
