+++ packages/tanstack-start/src/exports/server.ts
@@ -6,9 +6,15 @@ export { getRequestI18n } from '../utilities/getRequestI18n.server.js'
 export { getRequestLocale } from '../utilities/getRequestLocale.js'
 export { getRequestTheme } from '../utilities/getRequestTheme.js'
 export { handleGraphQL } from '../utilities/graphqlHandler.server.js'
+export { handleAPIRoute } from '../utilities/handleAPIRoute.js'
 export { handleServerFunctions } from '../utilities/handleServerFunctions.js'
 export { getImportMapOutputPath } from '../utilities/importMap.server.js'
 export { initReq } from '../utilities/initReq.server.js'
+export {
+  loadAdminPage,
+  type LoadAdminPageArgs,
+  type LoadAdminPageResult,
+} from '../utilities/loadAdminPage.js'
 export { type AdminPageMetadata, getAdminMeta } from '../utilities/meta.js'
 export { serializeForRsc } from '../utilities/serializeForRsc.js'
 export {
