+++ packages/tanstack-start/src/index.ts
@@ -2,3 +2,7 @@ export { TanStackComponentRenderer } from './elements/RenderComponent/index.js'
 export { TanStackRouterAdapter } from './elements/RouterAdapter/index.js'
 export { viteDevReloadStrategy } from './utilities/devReloadStrategy.js'
 export { type AdminPageMetadata, getAdminMeta } from './utilities/meta.js'
+export {
+  createServerFunctionClient,
+  stripUnserializable,
+} from './utilities/serverFunctionClient.js'
