+++ packages/tanstack-start/src/utilities/serverFunctionClient.ts
@@ -1,38 +1,9 @@
 import type { ServerFunctionClient, ServerFunctionClientArgs } from 'payload'
 
-import { createServerFn } from '@tanstack/react-start'
-
 /**
- * TanStack Start server function that dispatches to the shared Payload server
- * function registry (form-state, table-state, render-document, etc.) using the
- * RSC payload format — matching the Next.js adapter.
- *
- * Why a `createServerFn` instead of a hand-rolled `/api/server-function` JSON
- * route: server-rendered custom field components (returned by `buildFormState`
- * via `RenderServerComponent`) live as React elements inside the response. Raw
- * `JSON.stringify` strips them. TanStack Start's `createServerFn` wire format
- * uses seroval + an `$RSC` serialization adapter that serializes "RSC handles"
- * produced by `renderServerComponent` as Flight streams. The package's
- * `handleServerFunctions` runs the result through `serializeForRsc` to convert
- * React elements into those handles before they cross the wire.
- */
-export const runPayloadServerFn = createServerFn({ method: 'POST' })
-  .validator((args: ServerFunctionClientArgs): ServerFunctionClientArgs => args)
-  .handler(async ({ data }) => {
-    const { handleServerFunctions } = await import('@payloadcms/tanstack-start/server')
-    const config = (await import('@payload-config')).default
-    const { importMap } = await import('../importMap.js')
-
-    return (await handleServerFunctions({
-      name: data.name,
-      args: data.args,
-      config,
-      importMap,
-    })) as any
-  })
-
-/**
- * Client-side server function handler wired into `RootProvider.serverFunction`.
+ * Builds the client-side `ServerFunctionClient` wired into
+ * `RootProvider.serverFunction`, given the app's `createServerFn`-based
+ * dispatcher (`runServerFn`).
  *
  * Strips functions / symbols / RegExps / React elements from the args before
  * dispatching: the previous `fetch + JSON.stringify` pipeline silently dropped
@@ -41,18 +12,27 @@ export const runPayloadServerFn = createServerFn({ method: 'POST' })
  * `ServerFunctionsProvider`) that may pass the live form state — which can
  * carry stray functions — keep working without each call site sanitising.
  *
- * Delegates to `runPayloadServerFn`, so the response (including any RSC handles
- * for server-rendered custom components) is decoded by TanStack Start back into
+ * Delegates to `runServerFn`, so the response (including any RSC handles for
+ * server-rendered custom components) is decoded by TanStack Start back into
  * renderable React nodes on the client.
  */
-export const serverFunctionHandler: ServerFunctionClient = async (
-  args: ServerFunctionClientArgs,
-) => {
-  const safeArgs = stripUnserializable(args) as ServerFunctionClientArgs
-  return (await runPayloadServerFn({ data: safeArgs })) as ReturnType<ServerFunctionClient>
+export function createServerFunctionClient({
+  runServerFn,
+}: {
+  runServerFn: (input: { data: ServerFunctionClientArgs }) => Promise<unknown>
+}): ServerFunctionClient {
+  return (async (args: ServerFunctionClientArgs) => {
+    const safeArgs = stripUnserializable(args) as ServerFunctionClientArgs
+    return await runServerFn({ data: safeArgs })
+  }) as ServerFunctionClient
 }
 
-function stripUnserializable(
+/**
+ * Recursively removes values TanStack Start's seroval wire format cannot
+ * serialize (functions, symbols, RegExps, React elements, cyclic refs) and
+ * normalizes `Date` subclasses to plain `Date`.
+ */
+export function stripUnserializable(
   value: unknown,
   cache: WeakMap<object, unknown> = new WeakMap(),
   ancestors: WeakSet<object> = new WeakSet(),
