+++ test/admin-bar/app-tanstack/app/_payload/admin.index.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute, notFound, redirect } from '@tanstack/react-router'
 
 import { getAdminMeta } from '@payloadcms/tanstack-start'
-import { loadAdminPageRSC } from '~/functions/adminPageRSC.functions.js'
+import { loadAdminPageRSC } from './adminPageRSC.functions.js'
 
 export const Route = createFileRoute('/_payload/admin/')({
   // Mirror admin.$.tsx: surface query params in `loaderDeps` so changes like
