+++ test/admin-bar/app-tanstack/app/_payload/switchLanguage.functions.ts
@@ -1,0 +1,14 @@
+import { createServerFn } from '@tanstack/react-start'
+
+/**
+ * Thin TanStack Start binding for the language switcher. Logic lives in
+ * `switchLanguage` from `@payloadcms/tanstack-start`.
+ */
+export const switchLanguageFn = createServerFn({ method: 'POST' })
+  .validator((data: string) => data)
+  .handler(async ({ data: language }) => {
+    const { switchLanguage } = await import('@payloadcms/tanstack-start/server')
+    const config = (await import('@payload-config')).default
+
+    await switchLanguage({ config, language })
+  })
